requirements.txt
# Core dependencies for Parquet reading and S3 access
pyarrow>=14.0.0
s3fs>=2023.12.0

# PostgreSQL connectivity
psycopg2-binary>=2.9.9

# Trino connectivity (for validation)
trino>=0.328.0

# Data manipulation
pandas>=2.1.0
numpy>=1.24.0

# Configuration management
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Date/time utilities
python-dateutil>=2.8.2

# Testing and data generation
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
koku-nise>=5.3.0  # For generating synthetic OCP/AWS test data

# Logging and monitoring
structlog>=23.2.0
colorama>=0.4.6

# Performance profiling and monitoring
memory-profiler>=0.61.0
py-spy>=0.3.14
psutil>=5.9.0

# Type hints
types-PyYAML>=6.0.12

